Create README.md
The content of these CSS files can be used to customize the look and feel of a Blogger blog. It includes a variety of features, such as custom colors, fonts, and layouts. 
All style sheets are rendered correctly on all devices since I added @media rule to all relevant HTML elements.